Made Changes on the Landing Page
